Move authentication to a child process
If the wayland server's write buffer for a client connection fills up,
it terminates the client. Since authentication using PAM is blocking,
this can happen fairly easily if e.g. many input events are triggered
during authentication. As such, we must make this authentication
asynchronous.

It's also not bad to have a robust poll based event loop for future
extension.