auth: fix error handling edge case in conversation
PAM expects to be able to call free on every response struct with a
non-null resp field. However, we currently leave this field
undefined, which can cause an invalid free depending on how the PAM
modules are written.

This has been seen in practice with the pam_systemd_home module.