Lock: add -ignore-empty-password option
This allows users who, habitually or otherwise, use the Enter key to
wake their computer to avoid causing multiple failed authentication
attempts in the process.

Unfortunately, this doesn't seem viable as a default as there appear
to be valid configurations where submitting an empty password to PAM
is exactly what the user intends. To quote a commit message from when
i3lock added this feature:

> Some users may want to validate an empty password: PAM may rely on
> other sources to unlock the screen, like the presence of a token or
> the proximity of some Bluetooth device.