Lock: add -ready-fd option
My original idea was sending SIGUSR1, but ifreund suggested this, also
present in swaylock. The advantage of this approach is that it's simpler
and works well with at least two service managers (s6-rc, dinit).

Figuring out how to use this in a shell script took me longer than i'd
like to admit. This is what i ended up with:

    ( waylock-wrapper -ready-fd 3 3>&1 1>&2 & ) | read && suspend